Add files via upload
Business problem: You are a data analyst working for an online retail Superstore-type company that sells a wide variety of items.
You’re given a spreadsheet with raw data on your company’s sales data, and your job is to aggregate the data. The spreadsheet
contains the following information:

• Row ID
• Order ID
• Order Date
• Ship Date
• Ship Mode
• Segment
• State
• Region
• Product ID
• Category
• Sub-Category
• Product Name
• Sales
• Quantity
• Discount
• Profit

ADD COLUMNS IN THE SPREADSHEET TO ANSWER THE FOLLOWING QUESTIONS FOR EACH MONTH

1. What is our company’s total revenue?
2. What is our company’s total profit?
3. What is the sum of all order items that represent losses?
a. Hint: This is the sum of all negative numbers in the “Profit” column.
4. What is the most popular item?
5. What region had the highest sales?
6. What Category had the highest sales?
7. What is the most profitable item?
8. Create at least one plot that helps visually show the answer to one of the above questions.
9. (Optional) You are tasked with making suggestions for a marketing campaign. What region, categories, or products do you
think would make good marketing targets? Use data to support your answer.

Advance VBA
1. Open up this Excel Easy tutorial. Follow this tutorial to create a simple macro that types “Hello!” when you click
an in-sheet button.
2. Create a new macro to build a button called “Calculate Profit” that displays our company’s total profit.
3. Optional: What other questions might your boss be interested in? Is it better to produce these with VBA or simple formulas
and shortcuts? What’s an example of a question that requires VBA?